fix(metrics): pin the metric grid to uniform float64 (kj5 follow-up)
vbt's row_stack merge concats per-window metric frames; a window whose
every candidate finalizes a metric to None (e.g. sharpe/win_rate in a
no-trade window) produced an all-NA object column, tripping pandas'
deprecated all-NA dtype reconciliation (FutureWarning, behavior change
on a future pandas bump). Surfaced by kj5's un-masking regression test:
the all-invalid short-circuit frame is float64 and disagreed with such
object columns.

- central_metrics_from_grouped_accessors builds the grid with
  dtype="float64" (None -> NaN); the None contract stays at the
  optional_float seam (ranking/Evidence), so Evidence bytes are
  unchanged
- _nan_metric_frame documents the shared grid dtype contract
- the lightweight test pins the contract: NaN values inside a
  uniformly float64 grid, with a dtypes guard

+++ research/aegis_research/metrics/accessors.py
@@ -50,7 +50,14 @@ def central_metrics_from_grouped_accessors(
         rows.append(row)
 
     index = pd.MultiIndex.from_tuples(candidate_keys, names=param_names)
-    return pd.DataFrame(rows, index=index)
+    # The metric grid is uniformly float64 (None -> NaN). Without the explicit
+    # dtype, a column whose every candidate finalizes to None (e.g. sharpe in a
+    # no-trade window) lands as an all-NA object column, and pandas' deprecated
+    # all-NA dtype reconciliation fires (FutureWarning) when vbt row-stacks it
+    # with the float64 frames from other windows (incl. the runner's all-invalid
+    # NaN frame). Downstream consumers read values through optional_float, so
+    # NaN-vs-None inside the grid is invisible to ranking and Evidence.
+    return pd.DataFrame(rows, index=index, dtype="float64")
 
 
 def _ith(series: Any, idx: int) -> Any:
